requirements.txt
Django
djangorestframework
psycopg2
drf-spectacular
django-cors-headers
requests
Pillow
paho-mqtt>=1.6
